requirements.txt
art==5.1
autopep8==1.5.4
click==7.1.2
colorzero==1.1
Flask==1.1.2
gpiozero==1.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pigpio==1.78
pycodestyle==2.6.0
python-dateutil==2.8.1
PyYAML==5.3.1
RPi.GPIO==0.7.0
six==1.15.0
suntime==1.2.5
toml==0.10.2
Werkzeug==1.0.1
